scripts/syllabus_sync.py
import sys
import os
import requests
from bs4 import BeautifulSoup
from supabase import create_client, Client
from dotenv import load_dotenv

# Load credentials from the main project's .env.local
load_dotenv(os.path.join(os.getcwd(), '.env.local'))

SUPABASE_URL = os.environ.get("NEXT_PUBLIC_SUPABASE_URL")
SUPABASE_KEY = os.environ.get("NEXT_PUBLIC_SUPABASE_ANON_KEY")

if not SUPABASE_URL or not SUPABASE_KEY:
    print("Error: Supabase credentials not found in .env.local")
    sys.exit(1)

supabase: Client = create_client(SUPABASE_URL, SUPABASE_KEY)

def sync_branch_syllabus(scheme, branch_code, semester):
    """
    Scrapes and syncs syllabus data for a specific branch and semester.
    Note: VTU syllabus is often in PDFs. This script focuses on the 
    Subject Catalog synchronization from structured data or a source URL.
    """
    print(f"\n[SYLLABUS] Syncing {scheme} Scheme | {branch_code} | Sem {semester}...")
    
    # In a real scenario, you would scrape a URL like:
    # https://vturesource.com/vtu-subject-codes/{branch_code}-{scheme}-scheme.php
    
    # For now, this is a placeholder showing how to upsert to the database catalog
    # discovered via the Supabase Authority.
    
    # 1. Identify valid subjects (logic tailored to VTU naming convention)
    # This is where the 'Scraper' logic living in the backend would reside.
    pass

def run_importer(branch, scheme='2022'):
    """
    Runs a batch import for all 8 semesters of a branch.
    """
    print(f"=== Institutional Syllabus Importer: {branch} ({scheme}) ===")
    # Logic to populate the subject_catalog table
    # Once stored, the frontend calls this table instantly.
    print("Institutional Sync Complete.")

if __name__ == "__main__":
    if len(sys.argv) < 2:
        print("Usage: python syllabus_sync.py <BRANCH_CODE> [SCHEME]")
        sys.exit(1)
    
    branch = sys.argv[1].upper()
    sch = sys.argv[2] if len(sys.argv) > 2 else '2022'
    run_importer(branch, sch)
